deps/v8/test/webkit/fast/regex/dotstar-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

This page tests handling of parentheses subexpressions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS regexp1.exec('test') is null
PASS regexp1.exec('blah') is ['blah']
PASS regexp1.exec('1blah') is ['1blah']
PASS regexp1.exec('blah1') is ['blah1']
PASS regexp1.exec('blah blah blah') is ['blah blah blah']
PASS regexp1.exec('blah\nsecond') is ['blah']
PASS regexp1.exec('first\nblah') is ['blah']
PASS regexp1.exec('first\nblah\nthird') is ['blah']
PASS regexp1.exec('first\nblah2\nblah3') is ['blah2']
PASS regexp2.exec('test') is null
PASS regexp2.exec('blah') is ['blah']
PASS regexp2.exec('1blah') is ['1blah']
PASS regexp2.exec('blah1') is ['blah1']
PASS regexp2.exec('blah blah blah') is ['blah blah blah']
PASS regexp2.exec('blah\nsecond') is ['blah']
PASS regexp2.exec('first\nblah') is null
PASS regexp2.exec('first\nblah\nthird') is null
PASS regexp2.exec('first\nblah2\nblah3') is null
PASS regexp3.exec('test') is null
PASS regexp3.exec('blah') is ['blah']
PASS regexp3.exec('1blah') is ['1blah']
PASS regexp3.exec('blah1') is ['blah1']
PASS regexp3.exec('blah blah blah') is ['blah blah blah']
PASS regexp3.exec('blah\nsecond') is null
PASS regexp3.exec('first\nblah') is ['blah']
PASS regexp3.exec('first\nblah\nthird') is null
PASS regexp3.exec('first\nblah2\nblah3') is ['blah3']
PASS regexp4.exec('test') is null
PASS regexp4.exec('blah') is ['blah']
PASS regexp4.exec('1blah') is ['1blah']
PASS regexp4.exec('blah1') is ['blah1']
PASS regexp4.exec('blah blah blah') is ['blah blah blah']
PASS regexp4.exec('blah\nsecond') is null
PASS regexp4.exec('first\nblah') is null
PASS regexp4.exec('first\nblah\nthird') is null
PASS regexp4.exec('first\nblah2\nblah3') is null
PASS regexp5.exec('test') is null
PASS regexp5.exec('blah') is ['blah']
PASS regexp5.exec('1blah') is ['1blah']
PASS regexp5.exec('blah1') is ['blah1']
PASS regexp5.exec('blah blah blah') is ['blah blah blah']
PASS regexp5.exec('blah\nsecond') is ['blah']
PASS regexp5.exec('first\nblah') is ['blah']
PASS regexp5.exec('first\nblah\nthird') is ['blah']
PASS regexp5.exec('first\nblah2\nblah3') is ['blah2']
PASS regexp6.exec('test') is null
PASS regexp6.exec('blah') is ['blah']
PASS regexp6.exec('1blah') is ['1blah']
PASS regexp6.exec('blah1') is ['blah']
PASS regexp6.exec('blah blah blah') is ['blah blah blah']
PASS regexp6.exec('blah\nsecond') is ['blah']
PASS regexp6.exec('first\nblah') is ['blah']
PASS regexp6.exec('first\nblah\nthird') is ['blah']
PASS regexp6.exec('first\nblah2\nblah3') is ['blah']
PASS regexp7.exec('test') is null
PASS regexp7.exec('blah') is ['blah']
PASS regexp7.exec('1blah') is ['1blah']
PASS regexp7.exec('blah1') is ['blah1']
PASS regexp7.exec('blah blah blah') is ['blah blah blah']
PASS regexp7.exec('blah\nsecond') is null
PASS regexp7.exec('first\nblah') is null
PASS regexp7.exec('first\nblah\nthird') is null
PASS regexp7.exec('first\nblah2\nblah3') is null
PASS regexp8.exec('test') is null
PASS regexp8.exec('blah') is ['blah','']
PASS regexp8.exec('1blah') is ['1blah','1']
PASS regexp8.exec('blah1') is ['blah1','']
PASS regexp8.exec('blah blah blah') is ['blah blah blah','blah blah ']
PASS regexp8.exec('blah\nsecond') is null
PASS regexp8.exec('first\nblah') is null
PASS regexp8.exec('first\nblah\nthird') is null
PASS regexp8.exec('first\nblah2\nblah3') is null
PASS regexp9.exec('test') is null
PASS regexp9.exec('blah') is ['blah']
PASS regexp9.exec('1blah') is ['1blah']
PASS regexp9.exec('blah1') is ['blah1']
PASS regexp9.exec('blah blah blah') is ['blah blah blah']
PASS regexp9.exec('blah\nsecond') is ['blah']
PASS regexp9.exec('first\nblah') is ['blah']
PASS regexp9.exec('first\nblah\nthird') is ['blah']
PASS regexp9.exec('first\nblah2\nblah3') is ['blah2']
PASS regexp10.exec('test') is null
PASS regexp10.exec('blah') is ['blah']
PASS regexp10.exec('1blah') is ['1blah']
PASS regexp10.exec('blah1') is ['blah1']
PASS regexp10.exec('blah blah blah') is ['blah blah blah']
PASS regexp10.exec('blah\nsecond') is ['blah']
PASS regexp10.exec('first\nblah') is ['blah']
PASS regexp10.exec('first\nblah\nthird') is ['blah']
PASS regexp10.exec('first\nblah2\nblah3') is ['blah2']
PASS regexp11.exec('test') is null
PASS regexp11.exec('blah') is ['blah']
PASS regexp11.exec('1blah') is ['1blah']
PASS regexp11.exec('blah1') is ['blah1']
PASS regexp11.exec('blah blah blah') is ['blah blah blah']
PASS regexp11.exec('blah\nsecond') is null
PASS regexp11.exec('first\nblah') is ['blah']
PASS regexp11.exec('first\nblah\nthird') is null
PASS regexp11.exec('first\nblah2\nblah3') is ['blah3']
PASS regexp12.exec('test') is null
PASS regexp12.exec('blah') is ['blah']
PASS regexp12.exec('1blah') is ['1blah']
PASS regexp12.exec('blah1') is ['blah1']
PASS regexp12.exec('blah blah blah') is ['blah blah blah']
PASS regexp12.exec('blah\nsecond') is null
PASS regexp12.exec('first\nblah') is ['blah']
PASS regexp12.exec('first\nblah\nthird') is null
PASS regexp12.exec('first\nblah2\nblah3') is ['blah3']
PASS regexp13.exec('abc\n123') is ['abc\n123']
PASS successfullyParsed is true

TEST COMPLETE
